+++ packages/http/src/server/controller.ts
@@ -1,6 +1,6 @@
 'use strict'
 
-import { Application } from '@supercharge/contracts'
+import { Application, Logger } from '@supercharge/contracts'
 
 export class Controller {
   /**
@@ -15,5 +15,12 @@ export class Controller {
    */
   constructor (app: Application) {
     this.app = app
+  }
+
+  /**
+   * Returns the application logger instance.
+   */
+  protected logger (): Logger {
+    return this.app.make<Logger>('logger')
   }
 }